moved resume pages to component folder
